fix(webapp): resolve Next.js 16 build conflict by renaming local proxy helper
